move tests to directory root
