+++ .cargo/config.toml
@@ -1,4 +1,3 @@
 [alias]
-
 bl = "build --features daemon"
 ch = "check --features daemon"
