+++ boot-core/Cargo.toml
@@ -86,4 +86,3 @@ cw20-base               = { git = "https://github.com/Abstract-OS/cw-plus.git" }
 speculoos               = "0.9.0"
 ctor                    = "0.2.0"
 duct                    = "0.13"
-once_cell               = "1.8"
