+++ boot-core/src/daemon/mod.rs
@@ -5,9 +5,10 @@ pub mod error;
 pub mod networks;
 pub mod querier;
 mod sender;
+pub use sender::Wallet;
+pub mod json_file;
 pub mod state;
 pub mod tx_resp;
-pub mod utils;
 
 pub(crate) mod cosmos_modules {
     pub use cosmrs::proto::cosmos::auth::v1beta1 as auth;
