+++ boot-core/src/daemon/networks/archway.rs
@@ -1,4 +1,4 @@
-use crate::networks::{ChainInfo, ChainKind, NetworkInfo};
+use crate::networks::{ChainInfo, NetworkInfo, NetworkKind};
 
 pub const ARCHWAY_NETWORK: NetworkInfo = NetworkInfo {
     network_id: "archway",
