+++ boot-core/src/daemon/networks/mod.rs
@@ -6,7 +6,7 @@ pub mod neutron;
 pub mod osmosis;
 pub mod terra;
 
-pub use crate::daemon::state::{ChainInfo, ChainKind, NetworkInfo};
+pub use crate::daemon::state::{ChainInfo, NetworkInfo, NetworkKind};
 pub use archway::CONSTANTINE_1;
 pub use injective::{INJECTIVE_1, INJECTIVE_888};
 pub use juno::{JUNO_1, LOCAL_JUNO, UNI_6};
