+++ boot-core/src/lib.rs
@@ -8,7 +8,6 @@ mod interface;
 #[cfg(feature = "daemon")]
 mod keys;
 mod mock;
-// pub mod prelude;
 mod state;
 mod tx_handler;
 
@@ -34,12 +33,13 @@ pub use cw_multi_test::ContractWrapper;
 
 #[cfg(feature = "daemon")]
 pub use daemon::{
+    channel::DaemonChannel,
     core::{instantiate_daemon_env, Daemon},
     error::DaemonError,
     networks,
+    querier::DaemonQuerier,
     state::{DaemonOptions, DaemonOptionsBuilder},
-    // querier::DaemonQuerier,
-    // channel::DaemonChannel
+    Wallet,
 };
 
 #[cfg(feature = "daemon")]
@@ -57,5 +57,3 @@ impl<T: TxHandler + Clone> BootEnvironment for T {}
 /// Signals a supported execution environment for CosmWasm contracts
 pub trait CwEnv: TxHandler + Clone {}
 impl<T: TxHandler + Clone> CwEnv for T {}
-
-mod tests;
