--- boot-core/src/tests/daemon/mod.rs
@@ -1,6 +1,0 @@
-pub mod common;
-pub mod contract;
-pub mod general;
-pub mod grpc;
-pub mod querier;
-pub mod sender;
