--- boot-core/src/tests/mod.rs
@@ -1,2 +1,0 @@
-pub mod common;
-pub mod daemon;
